+++ CHANGES.rst
@@ -34,6 +34,8 @@ Changelog
   [chris-adam]
 - Added versions table in edit mode, removed useless columns and made clickable columns not clickable
   [chris-adam, sgeulette]
+- Added setting to define signers from templates.
+  [chris-adam]
 
 3.1.2 (2026-04-24)
 ------------------